buildscript: reorganize user-defined variables
The only part that is changed regularly inside buildscript are
the OpenWrt and package revisions and the selected packages.

Move them up and put them into dedicated variables so it is more
obvious what to change and easier to do so.

While at it, remove outdated COMPAT_VERSION comment from Gluon
package; we don't pull it from there anyway. Update comment on
feed definition syntax as well.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>
Reviewed-by: Robert Langhammer <rlanghammer@web.de>