fff-dhcp: add fallback dns server
If the user did not specify a dns server, set a fallback.  Since we
already have multiple servers available on an anycast address, this
address is chosen instead of as specific server.  This should avoid the
issue that the default server might deprecate suddenly.

A warning is still displayed to remind the user to conscientiously
choose a server and not rely on defaults.

Signed-off-by: Johannes Kimmel <fff@bareminimum.eu>
Reviewed-by: Fabian Bläse <fabian@blaese.de>