fff-layer3-config: generate default configuration
Generate a default layer3 configuration. The default configuration is
designed to replicate the state of all supported devices on their first
boot.

In an upcoming change, the first boot configuration shared between the
node and the layer3 variant is removed from the layer3 variant, so that
the gateway configuration is the only configuration source, and no
default state is assumed.

Signed-off-by: Fabian Bläse <fabian@blaese.de>