fff-network: remove cpuport from device definitions
Instead of explicitly adding the CPU port to the port lists of swconfig
devices, add it programmatically. A function to get the CPU port is
already present.

The configure-layer3 hooks already add the CPU port programmatically,
so this change aligns the behavior of the network.* files with the
behavior of configure-layer3. As a result, mistakes in the CPU port
function and the network.* port lists result in similar defects in both
firmware variants.

Signed-off-by: Fabian Bläse <fabian@blaese.de>