fff-network: Add temporary workaround for dsa bridge configuration
If the bridge created for DSA devices is configured without any members,
a single netifd reload is currently not sufficient to correctly apply
network settings. This is due to a bug in netifd (FS#4104).

Temporarily add a simple workaround by always configuring the bridge,
even if it does not have any member ports.

Signed-off-by: Fabian Bläse <fabian@blaese.de>
Reviewed-by: Johannes Kimmel <fff@bareminimum.eu>